Moves komodo-rest services into their own package
* Gives them an obvious home
